extract useResize hook
